rename the package io.sbk -> io.sbm
