+++ sbm/src/main/java/io/sbm/logger/SetRW.java
@@ -8,7 +8,7 @@
  * http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.sbk.logger;
+package io.sbm.logger;
 
 /**
  * Interface SetRW.
