+++ sbm/src/main/java/io/sbm/ram/impl/RamTotalWindowLatencyPeriodicRecorder.java
@@ -8,7 +8,7 @@
  * http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.sbk.ram.impl;
+package io.sbm.ram.impl;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.perl.api.LatencyRecord;
@@ -16,11 +16,11 @@ import io.perl.api.LatencyRecordWindow;
 import io.perl.logger.Print;
 import io.perl.api.ReportLatencies;
 import io.perl.api.impl.TotalLatencyRecordWindow;
+import io.sbm.ram.RamPeriodicRecorder;
 import io.sbp.grpc.LatenciesRecord;
 import io.sbk.logger.ReadRequestsLogger;
-import io.sbk.logger.SetRW;
+import io.sbm.logger.SetRW;
 import io.sbk.logger.WriteRequestsLogger;
-import io.sbk.ram.RamPeriodicRecorder;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
